requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-compressor==2.2
django-compressor-toolkit==0.6.0
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-ktag==1.0.3
django-npm==1.0.0
django-phonenumber-field==2.4.0
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
idna==2.8
phonenumbers==8.10.11
Pillow==6.0.0
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.1
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.0.12
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
